[REFACTOR] Certik Audit Recommendation - Exhibit 2
Description:
In general, underscore variables are used to denote private members of a contract. In this
instance, the underscore prefix is incorrectly utilized in a “public” variable to discern it from the
inputs of the “setPayoutWallet” function.

Recommendations:
In such instances, the underscore is instead used on the inputs of the function rather than the
members themselves. This is to conform to the camel-case function / variable naming
convention and do not allude to a variable being “private” due to its underscore prefix. We
advise that the underscore is swapped between the “address payoutWallet” input of
“setPayoutWallet” on L20 and the “address payable public _payoutWallet” on L14 of the
contract.